+++ chat-server/src/services/openAiChatLlm.ts
@@ -127,202 +127,3 @@ function validateOpenAiConversation(
     }
   }
 }
-
-// API CHAT
-export interface LlmFunction {
-  definition: FunctionDefinition;
-  function: (params: unknown) => unknown;
-  dynamic?: boolean;
-  name: string;
-}
-
-function makeApiChat(
-  openAiClient: OpenAIClient,
-  deploymentName: string,
-  systemPromptPersonality: string,
-  findApiSpecActionFunc: (params: unknown) => unknown
-) {
-  const baseSystemPrompt = `${systemPromptPersonality}
-
-Use the find_api_spec_action function to find an action in the API spec when the user asks you to perform an action.
-If none of the available functions match the user's query, use this function.
-
-Use the clear_api_spec_actions function to clear the API spec actions for the set of available actions.
-Do this when a user wants to perform a new action from the action
-`;
-  const baseFunctions: LlmFunction[] = [
-    {
-      name: "find_api_spec_action",
-      definition: {
-        name: "find_api_spec_action",
-        description: "Find an action in the API spec",
-        parameters: {
-          type: "object",
-          properties: {
-            query: { type: "string", description: "repeat the user's query" },
-          },
-        },
-      },
-      function: findApiSpecActionFunc,
-    },
-    {
-      name: "clear_api_spec_actions",
-      definition: {
-        name: "clear_api_spec_actions",
-        description:
-          "Clear the API spec actions for the set of available actions",
-        parameters: {
-          type: "object",
-          properties: {
-            clear: "boolean",
-          },
-        },
-      },
-      function: () => {
-        return baseFunctions;
-      },
-    },
-  ];
-
-  return {
-    async answerAwaited(
-      query: string,
-      messages: OpenAiChatMessage[],
-      availableFunctions: LlmFunction[],
-      options: GetChatCompletionsOptions,
-      requestApiFunc: (
-        pathParams: Record<string, unknown>,
-        queryParams: Record<string, unknown>,
-        body: Record<string, unknown>,
-        action: string
-      ) => unknown
-    ) {
-      const functions = availableFunctions.map((f) => f.definition);
-      const newMessages = [
-        {
-          role: "system",
-          content: makeSystemPrompt(baseSystemPrompt, availableFunctions),
-        },
-        ...messages.filter((message) => message.role !== "system"),
-        { role: "user", content: query },
-      ] satisfies OpenAiChatMessage[];
-      const messageOptions: GetChatCompletionsOptions = {
-        ...options,
-        functions: functions,
-        functionCall: "auto",
-      };
-      let responseAvailableFunctions = [...functions];
-      const response = await openAiClient.getChatCompletions(
-        deploymentName,
-        newMessages,
-        messageOptions
-      );
-      const { choices } = response;
-      const choice = choices[0];
-      if (choice.message?.functionCall) {
-        const { name } = choice.message.functionCall;
-        const functionToCall = availableFunctions.find(
-          (f) => f.definition.name === name
-        );
-        if (functionToCall && functionToCall.dynamic) {
-          const functionResponse = functionToCall.function(
-            choice.message.functionCall.arguments
-          );
-
-          newMessages.push({
-            role: "function",
-            content: JSON.stringify(functionResponse),
-          });
-        } else if (
-          functionToCall &&
-          functionToCall.name === "clear_api_spec_actions"
-        ) {
-          responseAvailableFunctions = baseFunctions;
-        } else if (
-          functionToCall &&
-          functionToCall.name === "find_api_spec_action"
-        ) {
-          // TODO: refactor this to have a function makeLlmFunction() that creates the LlmFunction based on the arguments
-          const apiSpecAction = functionToCall.function(
-            choice.message.functionCall.arguments
-          );
-          responseAvailableFunctions.push(apiSpecAction);
-        }
-        const responseBasedOnFunction = await openAiClient.getChatCompletions(
-          deploymentName,
-          newMessages,
-          messageOptions
-        );
-        newMessages.push(responseBasedOnFunction.choices[0].message);
-        return {
-          newMessages,
-          availableFunctions: responseAvailableFunctions,
-        };
-      }
-    },
-  };
-}
-
-// TODO
-function makeLlmFunction() {}
-
-function makeSystemPrompt(
-  baseSystemPrompt: string,
-  currentFunctions: LlmFunction[]
-) {
-  const dynamicFuncs = currentFunctions.filter((f) => f.dynamic);
-  if (dynamicFuncs.length > 0) {
-    return (
-      baseSystemPrompt +
-      `Call the functions ${dynamicFuncs
-        .map((func) => func.definition.name)
-        .join(
-          ", "
-        )} only when you have all necessary parameters to complete the action. If you don't yet have all the necessary information, continue asking the user for more information until you have it all.`
-    );
-  } else return baseSystemPrompt;
-}
-
-// Data Models
-
-export interface ApiPage {
-  /**
-    API spec
-   */
-  spec: Record<string, unknown>;
-  name: string;
-  url: string;
-}
-
-/**
-  Type for what we store in the DB
- */
-export interface ApiEmbeddedContent {
-  /**
-    definition of the endpoint from the API spec
-   */
-  definition: OpenApiEndpointDefinition;
-  /**
-    Function description from the API spec + metadata as frontmatter
-   */
-  embedding_text: string;
-  /**
-    Vector representation of the embedding_text.
-   */
-  embedding: number[];
-  /**
-    The API spec action name
-   */
-  action_name: string;
-  /**
-    Metadata. Used in the embedding text and also could be used for search.
-   */
-  metadata: Record<string, unknown>;
-}
-
-/**
-  Helper type for the OpenAPI endpoint definition...can probably find a type definition from a package for this.
- */
-interface OpenApiEndpointDefinition {
-  // TODO
-}
